add home component and react router v6
